% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions.R
\name{mrmt_nest}
\alias{mrmt_nest}
\title{Divisie un tableau selon les catégories}
\usage{
mrmt_nest(df)
}
\arguments{
\item{df}{Une base de donnée traitée avec \code{\link{mrmt_recode}}}
}
\value{
Un tibble avec les différentes modalités et sous-tableaux
}
\description{
Cette fonction va diviser la base de donnée d'entrée en différents sous-tableaux selon les catégories MRMT classiques
}
\examples{
mrmt_nest(ZPHH)
}
