% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions.R
\name{mrmt_recode}
\alias{mrmt_recode}
\title{Recode selon les catégories de l'OFS}
\usage{
mrmt_recode(df, sexe, age, act_prof, revenu, deg_urb)
}
\arguments{
\item{df}{Une base de donnée}

\item{sexe}{La colonne 'sexe' de la base de donnée choisie (nom en output : Sexe)}

\item{age}{La colonne 'âge' de la base de donnée choisie (nom en output : Âge)}

\item{act_prof}{La colonne 'activité professionnelle' de la base de donnée choisie (nom en output : Activité_professionnelle)}

\item{revenu}{La colonne 'revenu' de la base de donnée choisie (nom en output : Revenu_mensuel_du_ménage)}

\item{deg_urb}{La colonne 'degré d'urbanisation' de la base de donnée choisie (nom en output : Degré_d_urbanisation_du_domicile)}
}
\value{
La base de donnée d'entrée plus les nouvelles colonnes recodées selon la nomenclature OFS
}
\description{
Cette fonction va créer, à partir d'une base de donnée, de nouvelles colonnes correspondant aux catégories classiques MRMT (les lignes du tableau A2).
}
\examples{
mrmt_recode(ZPHH_raw,gesl,alter,ERWERB,F20601,W_staedt_char_2012)
}
