% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions.R
\name{mrmt_wmean_by}
\alias{mrmt_wmean_by}
\title{Moyennes pondérées pour une modalité spécifique sur un tableau nesté}
\usage{
mrmt_wmean_by(nestdf, mesure, category, value, weights)
}
\arguments{
\item{nestdf}{Un base de donnée traitée avec \code{\link{mrmt_nest}}}

\item{mesure}{La variable d'intérêt}

\item{category}{La catégorie voulue}

\item{value}{La modalité de la catégorie voulue}

\item{weights}{La colonne des poids}
}
\value{
Un tibble avec la moyenne pondérée, l'intervalle de confiance et le nombre d'individu de chaque sous-tableau
}
\description{
Cette fonction retourne la moyenne pondérée et l'intervalle de confiance d'une variable d'intérêt pour la modalité d'une catégorie voulue, ainsi que le nombre d'invidivu pour chaque sous-tableau
}
\examples{
mrmt_wmean_by(DéplacementZPHH_nest,déplacements,Motif,"Travail",WP) # moyenne pondérée du nombre de déplacement pour le motif travail
}
