% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions.R
\name{mrmt_wprop}
\alias{mrmt_wprop}
\title{Proportions pondérées sur un tableau nesté}
\usage{
mrmt_wprop(nestdf, category, value, weights)
}
\arguments{
\item{nestdf}{Un base de donnée traitée avec \code{\link{mrmt_nest}}}

\item{category}{La variable d'indérêt}

\item{value}{La modalité de la variable d'intérêt}

\item{weights}{La colonne des poids}
}
\value{
Un tibble avec la proportion pondérée, l'intervalle de confiance et le nombre d'individu de chaque sous-tableau
}
\description{
Cette fonction retourne la proportion pondérée et l'intervalle de confiance de la modalité d'une variable d'intérêt, ainsi que le nombre d'invidivu pour chaque sous-tableau
}
\examples{
mrmt_wprop(ZPHH_nest,f50200,1,WP) # personnes mobiles le jour de référence, 1 = oui
}
