% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMTpackage.R
\name{MRMTpackage}
\alias{MRMTpackage}
\title{Le package MRMT}
\description{
Ce package contient une série de fonctions utiles pour générer des colonnes (voire la totalité) du tableau standard A2 dans le cadre du MRMT.
}
\author{
Basile Huguenin-Virchaux \email{basile.huguenin@ne.ch}
}
